// src/types/asperaNodeApi.ts
export interface components {
    schemas: {
// ...
        error: {
            error?: components["schemas"]["errorBody"];
        };
        /** @description The contents of the error response object. */
        errorBody: {
            /** @description The HTTP error code. */
            code?: number;
            /** @description The HTTP error message. */
            reason?: string;
            /** @description A longer error description. */
            user_message?: string;
        };
// ...
        filesBrowseRequest: {
            /** @description The full pathname of the directory to browse. */
            path: string;
            /** @description Do not return information about the first number of files and directories. Default 0. */
            skip?: number;
            /** @description Return no more than "count" results.  It is the client’s responsibility to keep track of number of results returned. Default 100. */
            count?: number;
            /** @description Sort the results by "path", "type" (directory/file), "size_a"(bytes, ascending), "size_d" (bytes, descending), "mtime_a"  (modification date, ascending), or "mtime_d" (modification date, descending). */
            sort?: string;
            /** @description Specify if sort and basename are case sensitive (true) or not (false). Default: false. */
            case_sensitive?: boolean;
            filters?: components["schemas"]["filesBrowseFilters"];
        };
// ...
        filesBrowseResponse: {
            self?: components["schemas"]["filesBrowseMeta"];
            items?: components["schemas"]["filesBrowseMeta"][];
        };
        filesPageResponse: {
            items?: components["schemas"]["filesPageMeta"][];
        };
        /** @description JSON object that contains the metadata associated with the base path of the request. */
        filesBrowseMeta: {
            /**
             * @description The pathname of the file, directory, or symbolic link.
             * @example /Documents/Users/aspera
             */
            path?: string;
            /**
             * @description The base name of the path.
             * @example sample.txt
             */
            basename?: string;
            /**
             * @description The type of the path. Value is "file", "directory", or "symbolic_link".
             * @example file
             */
            type?: string;
            /**
             * @description The size of the file or symbolic link, in bytes. Directory size is reported as 0.
             * @example 3290353589
             */
            size?: number;
            /**
             * @description The item's modification time, in ISO format.
             * @example 2011-01-31 11:59:59
             */
            mtime?: string;
            permissions?: components["schemas"]["filesBrowsePermissions"][];
            target?: components["schemas"]["filesBrowseTarget"];
        };
// ...
        "info-get-200": {
            /** @description node */
            application?: string;
            /** @description The version of the IBM Aspera server software on the node. */
            version?: string;
            /** @description The current time on the node in ISO format, for example 2018-02-28T21:28:06Z. */
            current_time?: string;
            /** @description The expiration date of the server software license as YYYY-MM-DD. */
            license_expiration_date?: string;
            /** @description The maximum transfer rate allowed by the server license. */
            license_max_rate?: string;
            /** @description Operating system specifications. */
            os?: string;
            /** @description The status of the node's connection to aejd, the Aspera on Cloud Activity app reporting service. Value is "connected" or "disconnected". */
            aej_status?: string;
            /** @description If Sync reporting is enabled on the node. Value is "yes" or "no". */
            async_reporting?: string;
            /** @description If transfer activity reporting is enabled on the node. Value is "yes" or "no". */
            transfer_activity_reporting?: string;
            /** @description The transfer user associated with the authentication method. */
            transfer_user?: string;
            /** @description The docroot that is configured for the transfer user, if any. */
            docroot?: string;
            /**
             * @description The unique ID of the node, such as dc3239h5-f872-48b9-v5h9-849rn233725.
             * @example dc3239h5-f872-48b9-v5h9-849rn233725
             */
            node_id?: string;
            /**
             * @description The unique ID of the cluster, such as 88613fc0-4ad8-47ee-99f0-019122f11575. This value is generated even when the node is not part of a cluster.
             * @example 88613fc0-4ad8-47ee-99f0-019122f11575
             */
            cluster_id?: string;
            /**
             * @description An array of Watch Folder API ACLs that are associated with the user.
             * @example [
             *       "admin",
             *       "impersonation"
             *     ]
             */
            acls?: string[];
            access_key_configuration_capabilities?: components["schemas"]["ak_config_capabilities"];
            capabilities?: components["schemas"]["capabilities_items"][];
            settings?: components["schemas"]["settings_items"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}